requirements.txt
streamlit
feedparser
requests
yfinance
groq
python-dotenv
pandas
openpyxl
streamlit-autorefresh
google-generativeai
numpy
